#78 - Organized HTTP service production and tests classes by service
