+++ src/main/java/com/artipie/nuget/http/NuGet.java
@@ -33,7 +33,10 @@ import com.artipie.http.rq.RqMethod;
 import com.artipie.http.rs.RsStatus;
 import com.artipie.http.rs.RsWithStatus;
 import com.artipie.nuget.Repository;
+import com.artipie.nuget.http.content.PackageContent;
+import com.artipie.nuget.http.index.ServiceIndex;
 import com.artipie.nuget.http.metadata.PackageMetadata;
+import com.artipie.nuget.http.publish.PackagePublish;
 import java.net.URL;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
