+++ src/test/java/com/artipie/nuget/http/content/NuGetPackageContentTest.java
@@ -21,27 +21,21 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artipie.nuget.http;
+package com.artipie.nuget.http.content;
 
 import com.artipie.asto.Key;
 import com.artipie.asto.Storage;
 import com.artipie.asto.blocking.BlockingStorage;
 import com.artipie.asto.memory.InMemoryStorage;
-import com.artipie.http.Headers;
 import com.artipie.http.Response;
 import com.artipie.http.hm.RsHasBody;
 import com.artipie.http.hm.RsHasStatus;
 import com.artipie.http.rs.RsStatus;
-import com.google.common.io.Resources;
+import com.artipie.nuget.http.NuGet;
 import io.reactivex.Flowable;
-import java.io.ByteArrayOutputStream;
 import java.net.URL;
-import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.concurrent.CompletableFuture;
-import org.apache.http.HttpEntity;
-import org.apache.http.entity.mime.MultipartEntityBuilder;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.hamcrest.core.AllOf;
@@ -50,12 +44,13 @@ import org.junit.jupiter.api.Test;
 
 /**
  * Tests for {@link NuGet}.
+ * Package Content resource.
  *
  * @since 0.1
  * @checkstyle ClassDataAbstractionCouplingCheck (2 lines)
  */
 @SuppressWarnings("PMD.TooManyMethods")
-class NuGetTest {
+class NuGetPackageContentTest {
 
     /**
      * Storage used in tests.
@@ -138,56 +133,6 @@ class NuGetTest {
     }
 
     @Test
-    void shouldPutPackagePublish() throws Exception {
-        final Response response = this.putPackage(nupkg());
-        MatcherAssert.assertThat(
-            response,
-            new RsHasStatus(RsStatus.CREATED)
-        );
-    }
-
-    @Test
-    void shouldFailPutPackage() throws Exception {
-        MatcherAssert.assertThat(
-            "Should fail to add package which is not a ZIP archive",
-            this.putPackage("not a zip".getBytes()),
-            new RsHasStatus(RsStatus.BAD_REQUEST)
-        );
-    }
-
-    @Test
-    void shouldFailPutSamePackage() throws Exception {
-        this.putPackage(nupkg()).send(
-            (status, headers, body) -> CompletableFuture.allOf()
-        ).toCompletableFuture().join();
-        MatcherAssert.assertThat(
-            "Should fail to add same package when it is already present in the repository",
-            this.putPackage(nupkg()),
-            new RsHasStatus(RsStatus.CONFLICT)
-        );
-    }
-
-    @Test
-    void shouldFailGetPackagePublishFromNotBasePath() {
-        final Response response = this.nuget.response(
-            "GET /not-base/package HTTP/1.1",
-            Collections.emptyList(),
-            Flowable.empty()
-        );
-        MatcherAssert.assertThat(response, new RsHasStatus(RsStatus.NOT_FOUND));
-    }
-
-    @Test
-    void shouldFailGetPackagePublish() {
-        final Response response = this.nuget.response(
-            "GET /base/package HTTP/1.1",
-            Collections.emptyList(),
-            Flowable.empty()
-        );
-        MatcherAssert.assertThat(response, new RsHasStatus(RsStatus.METHOD_NOT_ALLOWED));
-    }
-
-    @Test
     void shouldGetPackageVersions() {
         final byte[] data = "example".getBytes();
         new BlockingStorage(this.storage).save(
@@ -217,26 +162,5 @@ class NuGetTest {
             ),
             new RsHasStatus(RsStatus.NOT_FOUND)
         );
-    }
-
-    private Response putPackage(final byte[] pack) throws Exception {
-        final HttpEntity entity = MultipartEntityBuilder.create()
-            .addBinaryBody("package.nupkg", pack)
-            .build();
-        final ByteArrayOutputStream sink = new ByteArrayOutputStream();
-        entity.writeTo(sink);
-        return this.nuget.response(
-            "PUT /base/package HTTP/1.1",
-            new Headers.From(
-                "Content-Type", entity.getContentType().getValue()
-            ),
-            Flowable.fromArray(ByteBuffer.wrap(sink.toByteArray()))
-        );
-    }
-
-    private static byte[] nupkg() throws Exception {
-        final URL resource = Thread.currentThread().getContextClassLoader()
-            .getResource("newtonsoft.json/12.0.3/newtonsoft.json.12.0.3.nupkg");
-        return Resources.toByteArray(resource);
     }
 }
